Refactored AuthController.php
