+++ src/Controller/LoginController.php
@@ -1,0 +1,77 @@
+<?php
+
+namespace App\Controller;
+
+use App\Mail\SymfonyMailer;
+use App\Repository\UserRepository;
+use App\Security\TokenGenerator;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
+
+
+class LoginController extends AbstractController
+{
+    /**
+     * @var EntityManagerInterface
+     */
+    //pour modifier la base de donnée
+    private $entityManager;
+
+    /**
+     * @var UserPasswordEncoderInterface
+     */
+    private $passwordEncoder;
+    /**
+     * @var SymfonyMailer
+     */
+    private $mailer;
+    /**
+     * @var TokenGenerator
+     */
+    private $tokenGenerator;
+
+    /**
+     * LoginController constructor.
+     * @param EntityManagerInterface $entityManager
+     * @param SymfonyMailer $mailer
+     * @param UserPasswordEncoderInterface $passwordEncoder
+     * @param TokenGenerator $tokenGenerator
+     */
+    public function __construct(
+        EntityManagerInterface $entityManager,
+        SymfonyMailer $mailer,
+        UserPasswordEncoderInterface $passwordEncoder,
+        TokenGenerator $tokenGenerator
+    )
+    {
+        $this->entityManager = $entityManager;
+        $this->mailer = $mailer;
+        $this->passwordEncoder = $passwordEncoder;
+        $this->tokenGenerator = $tokenGenerator;
+    }
+
+    /**
+     * @Route("/api/login_check", name="login")
+     * @param Request $request
+     * @param UserRepository $repository
+     * @return JsonResponse
+     */
+    public function login(Request $request, UserRepository $repository): JsonResponse
+    {
+        $data = json_decode($request->getContent(), true);
+        $user = $repository->findOneBy(['email' => $data["username"]]);
+        if ($user === null){
+            return $this->json('user not identified', 400);
+        }
+        return $this->json([
+            'username'=>$user->getUsername(),
+            'roles'=>$user->getRoles(),
+            'login'=>$user->getLogin()
+        ]);
+    }
+
+}
